Day 1 Improved Solution
Solution for Rearrange Geek and his Classmates
Each of the N students in the class can be identified by a unique roll number X and each
desk has a number i associated with it. Only one student can sit on one desk.
Each time the teacher turns her back, a student with roll number X sitting on desk number i
gets up and takes the place of the student with roll number i.
If the current position of N students in the class is given to you in an array, such that i is the
 desk number and a[i] is the roll number of the student sitting on the desk, can you modify
//  a[ ] to represent the next position of all the students ?