Day 7 Solution
Solution for the problem
Given an array of size N, find the number of distinct pairs {a[i], a[j]} (i != j) in the array with their sum greater than 0.
Your Task:
You don't need to read input or print anything. Complete the function ValidPair() which takes the array a[ ] and size of array N as input parameters and returns the count of such pairs.
Expected Time Complexity: O(N * Log(N))
Expected Auxiliary Space: O(1)