Day 13 Problem Check Tree Traversal of 30 Days of Code
Solution for problem
Given Preorder, Inorder and Postorder traversals of some tree of size N. The task is to check if they are all of the same tree or not.
function checktree() which takes the array preorder[ ], inorder[ ], postorder[ ] and integer N as input parameters and returns true if the three traversals are of the same tree or not.

Expected Time Complexity: O(N2)
Expected Auxiliary Space: O(N)